+++ miden-tx/src/tests/kernel_tests/test_faucet.rs
@@ -12,7 +12,7 @@ use miden_lib::{
 use miden_objects::{
     accounts::account_id::testing::{
         ACCOUNT_ID_FUNGIBLE_FAUCET_ON_CHAIN, ACCOUNT_ID_FUNGIBLE_FAUCET_ON_CHAIN_1,
-        ACCOUNT_ID_NON_FUNGIBLE_FAUCET_ON_CHAIN, ACCOUNT_ID_NON_FUNGIBLE_FAUCET_ON_CHAIN_1,
+        ACCOUNT_ID_NON_FUNGIBLE_FAUCET_ON_CHAIN_1,
     },
     assets::{FungibleAsset, NonFungibleAsset},
     testing::{
@@ -161,14 +161,13 @@ fn test_mint_fungible_asset_fails_saturate_max_amount() {
 #[test]
 fn test_mint_non_fungible_asset_succeeds() {
     let tx_context = TransactionContextBuilder::with_non_fungible_faucet(
-        ACCOUNT_ID_NON_FUNGIBLE_FAUCET_ON_CHAIN,
+        NonFungibleAsset::mock_issuer().into(),
         ONE,
         false,
     )
     .build();
 
-    let non_fungible_asset =
-        NonFungibleAsset::mock(ACCOUNT_ID_NON_FUNGIBLE_FAUCET_ON_CHAIN, &NON_FUNGIBLE_ASSET_DATA);
+    let non_fungible_asset = NonFungibleAsset::mock(&NON_FUNGIBLE_ASSET_DATA);
 
     let code = format!(
         "
@@ -215,8 +214,7 @@ fn test_mint_non_fungible_asset_succeeds() {
 fn test_mint_non_fungible_asset_fails_not_faucet_account() {
     let tx_context = TransactionContextBuilder::with_standard_account(ONE).build();
 
-    let non_fungible_asset =
-        NonFungibleAsset::mock(ACCOUNT_ID_NON_FUNGIBLE_FAUCET_ON_CHAIN, &[1, 2, 3, 4]);
+    let non_fungible_asset = NonFungibleAsset::mock(&[1, 2, 3, 4]);
 
     let code = format!(
         "
@@ -241,8 +239,7 @@ fn test_mint_non_fungible_asset_fails_not_faucet_account() {
 fn test_mint_non_fungible_asset_fails_inconsistent_faucet_id() {
     let tx_context = TransactionContextBuilder::with_standard_account(ONE).build();
 
-    let non_fungible_asset =
-        NonFungibleAsset::mock(ACCOUNT_ID_NON_FUNGIBLE_FAUCET_ON_CHAIN_1, &[1, 2, 3, 4]);
+    let non_fungible_asset = NonFungibleAsset::mock(&[1, 2, 3, 4]);
 
     let code = format!(
         "
@@ -266,14 +263,13 @@ fn test_mint_non_fungible_asset_fails_inconsistent_faucet_id() {
 #[test]
 fn test_mint_non_fungible_asset_fails_asset_already_exists() {
     let tx_context = TransactionContextBuilder::with_non_fungible_faucet(
-        ACCOUNT_ID_NON_FUNGIBLE_FAUCET_ON_CHAIN,
+        NonFungibleAsset::mock_issuer().into(),
         ONE,
         false,
     )
     .build();
 
-    let non_fungible_asset =
-        NonFungibleAsset::mock(ACCOUNT_ID_NON_FUNGIBLE_FAUCET_ON_CHAIN, &NON_FUNGIBLE_ASSET_DATA_2);
+    let non_fungible_asset = NonFungibleAsset::mock(&NON_FUNGIBLE_ASSET_DATA_2);
 
     let code = format!(
         "
@@ -433,14 +429,13 @@ fn test_burn_fungible_asset_insufficient_input_amount() {
 #[test]
 fn test_burn_non_fungible_asset_succeeds() {
     let tx_context = TransactionContextBuilder::with_non_fungible_faucet(
-        ACCOUNT_ID_NON_FUNGIBLE_FAUCET_ON_CHAIN,
+        NonFungibleAsset::mock_issuer().into(),
         ONE,
         false,
     )
     .build();
 
-    let non_fungible_asset_burnt =
-        NonFungibleAsset::mock(ACCOUNT_ID_NON_FUNGIBLE_FAUCET_ON_CHAIN, &[1, 2, 3]);
+    let non_fungible_asset_burnt = NonFungibleAsset::mock(&[1, 2, 3]);
 
     let code = format!(
         "
@@ -486,14 +481,13 @@ fn test_burn_non_fungible_asset_succeeds() {
 #[test]
 fn test_burn_non_fungible_asset_fails_does_not_exist() {
     let tx_context = TransactionContextBuilder::with_non_fungible_faucet(
-        ACCOUNT_ID_NON_FUNGIBLE_FAUCET_ON_CHAIN,
+        NonFungibleAsset::mock_issuer().into(),
         ONE,
         false,
     )
     .build();
 
-    let non_fungible_asset_burnt =
-        NonFungibleAsset::mock(ACCOUNT_ID_NON_FUNGIBLE_FAUCET_ON_CHAIN, &[1, 2, 3]);
+    let non_fungible_asset_burnt = NonFungibleAsset::mock(&[1, 2, 3]);
 
     let code = format!(
         "
@@ -524,8 +518,7 @@ fn test_burn_non_fungible_asset_fails_does_not_exist() {
 fn test_burn_non_fungible_asset_fails_not_faucet_account() {
     let tx_context = TransactionContextBuilder::with_standard_account(ONE).build();
 
-    let non_fungible_asset_burnt =
-        NonFungibleAsset::mock(ACCOUNT_ID_NON_FUNGIBLE_FAUCET_ON_CHAIN, &[1, 2, 3]);
+    let non_fungible_asset_burnt = NonFungibleAsset::mock(&[1, 2, 3]);
 
     let code = format!(
         "
@@ -556,11 +549,11 @@ fn test_burn_non_fungible_asset_fails_not_faucet_account() {
 
 #[test]
 fn test_burn_non_fungible_asset_fails_inconsistent_faucet_id() {
-    let non_fungible_asset_burnt =
-        NonFungibleAsset::mock(ACCOUNT_ID_NON_FUNGIBLE_FAUCET_ON_CHAIN_1, &[1, 2, 3]);
+    let non_fungible_asset_burnt = NonFungibleAsset::mock(&[1, 2, 3]);
 
+    // Run code from a different non-fungible asset issuer
     let tx_context = TransactionContextBuilder::with_non_fungible_faucet(
-        ACCOUNT_ID_NON_FUNGIBLE_FAUCET_ON_CHAIN,
+        ACCOUNT_ID_NON_FUNGIBLE_FAUCET_ON_CHAIN_1,
         ONE,
         false,
     )